+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 # Changelog
 
+## [Unreleased]
+### Added
+- so-repository modify command.
+
+### Fixed
+
+- so-repository ls command.
+
 ## [0.6.0] - 2022-12-13
 ### Added
 - Option for show or hide archived tasks.